Condense the evtx plugin to one file
