+++ default.db
@@ -1,0 +1,4 @@
+{
+    "generator": "recover-evtx",
+    "version": 1
+}
