--- volatility/EVTXtract/scripts/evtxview.py
@@ -1,1 +1,0 @@
-__author__ = 'willi'
